recipes-kernel/linux-toradex: disable spdif interfaces
Disable spdif interfaces in imx6q apalis evaluation board.
By default, pulseaudio detects both spdif and i2s interfaces as
`Built-in stereo`. As Jami selects its sources and sinks by name, this
creates a confusion between the devices. This makes it impossible to
select the i2s interface, as pulseaudio returns the spdif first.
As spdif is not used in our case, we simply disable it. To solve the
root cause of the problem, further work would be required on Jami, both
on daemon and clients.